examples/basic/python/tools/requirements.txt
shapesinc
python-dotenv